requirements/development.txt
-r common.txt
Flask-DebugToolbar==0.10.0
Flask-Testing==0.4.2
nose==1.3.0
